authy_v4_open/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
PyNaCl==1.5.0
python-dotenv==1.0.1
